# Improve `vec_compare()`