--- a/src/compare.c
+++ b/src/compare.c
@@ -148,6 +148,8 @@ int compare_scalar(SEXP x, R_len_t i, SEXP y, R_len_t j, bool na_equal) {
 
 // -----------------------------------------------------------------------------
 
+SEXP vec_compare_col(SEXP x, SEXP y, bool na_equal, SEXP info, R_len_t n);
+
 #define COMPARE(CTYPE, CONST_DEREF, SCALAR_COMPARE)     \
 do {                                                    \
   const CTYPE* xp = CONST_DEREF(x);                     \
@@ -159,18 +161,41 @@ do {                                                    \
 }                                                       \
 while (0)
 
-#define COMPARE_DF(SCALAR_COMPARE)                                     \
-do {                                                                   \
-  int n_col = Rf_length(x);                                            \
-                                                                       \
-  if (n_col != Rf_length(y)) {                                         \
-    stop_not_comparable(x, y, "must have the same number of columns"); \
-  }                                                                    \
-                                                                       \
-  for (R_len_t i = 0; i < n; ++i) {                                    \
-    p[i] = SCALAR_COMPARE(x, i, y, i, na_equal, n_col);                \
-  }                                                                    \
-}                                                                      \
+#define COMPARE_DF()                                                       \
+do {                                                                       \
+  int n_col = Rf_length(x);                                                \
+                                                                           \
+  if (n_col != Rf_length(y)) {                                             \
+    stop_not_comparable(x, y, "must have the same number of columns");     \
+  }                                                                        \
+                                                                           \
+  SEXP x_col;                                                              \
+  SEXP y_col;                                                              \
+                                                                           \
+  SEXP info = PROTECT(Rf_allocVector(VECSXP, 3));                          \
+  SEXP skip = PROTECT(Rf_allocVector(LGLSXP, n));                          \
+  SEXP count = PROTECT(Rf_allocVector(INTSXP, 1));                         \
+  INTEGER(count)[0] = n;                                                   \
+                                                                           \
+  int* p_skip = LOGICAL(skip);                                             \
+  memset(p_skip, 0, n * sizeof(int));                                      \
+                                                                           \
+  memset(p, 0, n * sizeof(int));                                           \
+                                                                           \
+  SET_VECTOR_ELT(info, 0, out);                                            \
+  SET_VECTOR_ELT(info, 1, skip);                                           \
+  SET_VECTOR_ELT(info, 2, count);                                          \
+                                                                           \
+  for (R_len_t i = 0; i < n_col; ++i) {                                    \
+    x_col = VECTOR_ELT(x, i);                                              \
+    y_col = VECTOR_ELT(y, i);                                              \
+                                                                           \
+    info = vec_compare_col(x_col, y_col, na_equal, info, n);               \
+  }                                                                        \
+                                                                           \
+  out = VECTOR_ELT(info, 0);                                               \
+  UNPROTECT(3);                                                            \
+}                                                                          \
 while (0)
 
 
@@ -193,7 +218,7 @@ SEXP vctrs_compare(SEXP x, SEXP y, SEXP na_equal_) {
   case vctrs_type_integer:   COMPARE(int, INTEGER_RO, int_compare_scalar); break;
   case vctrs_type_double:    COMPARE(double, REAL_RO, dbl_compare_scalar); break;
   case vctrs_type_character: COMPARE(SEXP, STRING_PTR_RO, chr_compare_scalar); break;
-  case vctrs_type_dataframe: COMPARE_DF(df_compare_scalar); break;
+  case vctrs_type_dataframe: COMPARE_DF(); break;
   case vctrs_type_scalar:    Rf_errorcall(R_NilValue, "Can't compare scalars with `vctrs_compare()`");
   case vctrs_type_list:      Rf_errorcall(R_NilValue, "Can't compare lists with `vctrs_compare()`");
   default:                   Rf_error("Unimplemented type in `vctrs_compare()`");
@@ -201,6 +226,82 @@ SEXP vctrs_compare(SEXP x, SEXP y, SEXP na_equal_) {
 
   UNPROTECT(1);
   return out;
+}
+
+#define COMPARE_COL(CTYPE, CONST_DEREF, SCALAR_COMPARE)     \
+do {                                                        \
+  const CTYPE* p_x = CONST_DEREF(x);                        \
+  const CTYPE* p_y = CONST_DEREF(y);                        \
+                                                            \
+  int cmp;                                                  \
+                                                            \
+  for (R_len_t i = 0; i < n; ++i, ++p_skip, ++p_x, ++p_y) { \
+    if (*p_skip) {                                          \
+      continue;                                             \
+    }                                                       \
+                                                            \
+    cmp = SCALAR_COMPARE(p_x, p_y, na_equal);               \
+                                                            \
+    if (cmp != 0) {                                         \
+      p_out[i] = cmp;                                       \
+      *p_skip = true;                                       \
+      --(*p_count);                                         \
+                                                            \
+      if (*p_count == 0) {                                  \
+        break;                                              \
+      }                                                     \
+    }                                                       \
+  }                                                         \
+}                                                           \
+while (0)
+
+#define COMPARE_DF_COL()                                       \
+do {                                                           \
+  int n_col = Rf_length(x);                                    \
+                                                               \
+  SEXP x_col;                                                  \
+  SEXP y_col;                                                  \
+                                                               \
+  for (R_len_t i = 0; i < n_col; ++i) {                        \
+    x_col = VECTOR_ELT(x, i);                                  \
+    y_col = VECTOR_ELT(y, i);                                  \
+                                                               \
+    info = vec_compare_col(x_col, y_col, na_equal, info, n);   \
+  }                                                            \
+                                                               \
+  return info;                                                 \
+}                                                              \
+while (0)
+
+SEXP vec_compare_col(SEXP x, SEXP y, bool na_equal, SEXP info, R_len_t n) {
+  SEXP out = VECTOR_ELT(info, 0);
+  SEXP skip = VECTOR_ELT(info, 1);
+  SEXP count = VECTOR_ELT(info, 2);
+
+  int* p_out = INTEGER(out);
+  int* p_skip = LOGICAL(skip);
+  int* p_count = INTEGER(count);
+
+  if (*p_count == 0) {
+    return info;
+  }
+
+  switch (vec_proxy_typeof(x)) {
+  case vctrs_type_logical:   COMPARE_COL(int, LOGICAL_RO, lgl_compare_scalar); break;
+  case vctrs_type_integer:   COMPARE_COL(int, INTEGER_RO, int_compare_scalar); break;
+  case vctrs_type_double:    COMPARE_COL(double, REAL_RO, dbl_compare_scalar); break;
+  case vctrs_type_character: COMPARE_COL(SEXP, STRING_PTR_RO, chr_compare_scalar); break;
+  case vctrs_type_dataframe: COMPARE_DF_COL();
+  case vctrs_type_scalar:    Rf_errorcall(R_NilValue, "Can't compare scalars with `vctrs_compare()`");
+  case vctrs_type_list:      Rf_errorcall(R_NilValue, "Can't compare lists with `vctrs_compare()`");
+  default:                   Rf_error("Unimplemented type in `vctrs_compare()`");
+  }
+
+  SET_VECTOR_ELT(info, 0, out);
+  SET_VECTOR_ELT(info, 1, skip);
+  SET_VECTOR_ELT(info, 2, count);
+
+  return info;
 }
 
 #undef COMPARE