--- a/src/compare.c
+++ b/src/compare.c
@@ -258,7 +258,7 @@ static SEXP df_compare_impl(SEXP x, SEXP y, bool na_equal, SEXP info, R_len_t n)
 
   SEXP x_col;
   SEXP y_col;
-  SEXP count;
+  int* p_count;
 
   for (R_len_t i = 0; i < n_col; ++i) {
     x_col = VECTOR_ELT(x, i);
@@ -267,8 +267,8 @@ static SEXP df_compare_impl(SEXP x, SEXP y, bool na_equal, SEXP info, R_len_t n)
     info = vec_compare_col(x_col, y_col, na_equal, info, n);
 
     // If we know all comparison values, break
-    count = VECTOR_ELT(info, 2);
-    if (*INTEGER(count) == 0) {
+    p_count = INTEGER(VECTOR_ELT(info, 2));
+    if (*p_count == 0) {
       break;
     }
   }