… tests/testthat/test-compare.R