Closes #648 

Still working on this, but incredibly promising results with a rewrite of `vec_compare()`. 

It follows the ideas of `vec_equal()` for the most part. But importantly I'm also experimenting with the idea of doing the comparisons column wise on data frames. This is way more efficient than row wise and I think we might be able to port this idea over to `vec_equal()` too.

The way it generally works is that we construct a logical `skip` vector initialized to all false values and with size equal to the number of rows in the data frame. As we go through the data frame columns, we flip the `skip` values to `TRUE` when we "know" the comparison value (i.e. if it is less than or greater than) and we don't ever evaluate that row again when comparing downstream columns. If all `skip` values are ever `TRUE` (as measured by a decreasing `count` that hits `0` when we know everything), we exit immediately.

Before

``` r
library(vctrs)
library(nycflights13)

bench::mark(
  vec_compare(flights$year, flights$year)
)
#> # A tibble: 1 x 6
#>   expression                                 min median `itr/sec` mem_alloc
#>   <bch:expr>                              <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_compare(flights$year, flights$year) 5.26ms 6.48ms      156.     135MB
#> # … with 1 more variable: `gc/sec` <dbl>

bench::mark(
  vec_compare(flights, flights)
)
#> # A tibble: 1 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                    <bch:t> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_compare(flights, flights)   189ms  197ms      5.07    58.2MB     2.54

flights_sort <- dplyr::arrange(flights, dep_time)

bench::mark(
  vec_compare(flights, flights_sort)
)
#> # A tibble: 1 x 6
#>   expression                            min median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_compare(flights, flights_sort) 39.1ms 42.3ms      23.9    11.6MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

After

``` r
library(vctrs)
library(nycflights13)

# vector speed up
bench::mark(
  vec_compare(flights$year, flights$year)
)
#> # A tibble: 1 x 6
#>   expression                                min median `itr/sec` mem_alloc
#>   <bch:expr>                              <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_compare(flights$year, flights$year) 351µs  509µs     1700.     133MB
#> # … with 1 more variable: `gc/sec` <dbl>

# all equal "worst case"
bench::mark(
  vec_compare(flights, flights)
)
#> # A tibble: 1 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                    <bch:t> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_compare(flights, flights)  12.3ms   13ms      75.6    2.92MB     2.04

flights_sort <- dplyr::arrange(flights, dep_time)

# not all equal, we compare column wise and skip if we already
# know the result
bench::mark(
  vec_compare(flights, flights_sort)
)
#> # A tibble: 1 x 6
#>   expression                            min median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_compare(flights, flights_sort) 2.52ms 3.13ms      306.    2.57MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

<sup>Created on 2019-11-06 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>