@lionel- 2019-11-07
> I'm also experimenting with the idea of doing the comparisons column wise on data frames. This is way more efficient than row wise and I think we might be able to port this idea over to vec_equal() 

yup we definitely need to finish vectorising data frame operations in vctrs, would be great to reach this point!