@DavisVaughan 2019-11-07
@lionel- ready for review

- Added encoding support for `vec_compare()`, which I had forgotten about when I did it for `vec_equal()`
- We validate `na_equal` now
- Added various tests to flesh out existing coverage

There is a big drop in test coverage for this file because `compare_scalar()` is no longer used by anything. This means it has no tests, along with `df_compare_scalar()`, which is only ever used by `compare_scalar()`. We include `compare_scalar()` in the vctrs.h api file (like `equal_scalar()`), but we don't actually use it.

Not sure what to do about that. It seems like it might someday be useful internally, but I don't really feel like it would be worth it to export it to the R side just for testing now.